Undid some bad deletes, reimplemented Steve's REST interface, bugfixes

+++ code/mobile/smartFridge/src/edu/rit/smartFridge/util/DataConnect.java
@@ -4,7 +4,6 @@ import java.io.Serializable;
 import java.util.Date;
 import java.util.List;
 
-import android.content.Context;
 import edu.rit.smartFridge.model.InventoryItem;
 import edu.rit.smartFridge.model.ShoppingList;
 
@@ -12,7 +11,7 @@ public interface DataConnect extends Serializable
 {
 	List<ShoppingList> getLists();
 
-	ShoppingList getList(int listId);
+	ShoppingList getList(long listId);
 
 	ShoppingList populateItems(ShoppingList list);
